+++ Modules/KiwiEngine/KiwiTypedObjects.hpp
@@ -40,7 +40,7 @@ namespace kiwi
             
             ObjectPlus(std::vector<Atom> const& args);
             
-            void receive(uint32_t index, std::vector<Atom> const& args) override;
+            void receive(size_t index, std::vector<Atom> const& args) override;
             
             void bang();
             
@@ -61,7 +61,7 @@ namespace kiwi
             
             ObjectPrint(std::vector<Atom> const& args);
             
-            void receive(uint32_t, std::vector<Atom> const& args) override;
+            void receive(size_t, std::vector<Atom> const& args) override;
             
         private:
             std::string m_name;
